Configure serialization of Instant
