+++ java/core/src/main/java/sleeper/core/statestore/FileReferenceSerDe.java
@@ -17,14 +17,15 @@
 package sleeper.core.statestore;
 
 import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
 import com.google.gson.reflect.TypeToken;
+
+import sleeper.core.util.GsonConfig;
 
 import java.util.Collection;
 import java.util.List;
 
 public class FileReferenceSerDe {
-    private final Gson gson = new GsonBuilder().create();
+    private final Gson gson = GsonConfig.standardBuilder().create();
 
     public String toJson(FileReference file) {
         return gson.toJson(file);
